Return object with commands and more
In this commit we've changed it so that Atheme::Service.command now returns an object.

Example:

  Atheme::ChanServ.info '#'
  => #<Atheme::ChanServ ...

Additionally we've added parsing of output to add additional methods to some commands.

Three were added in this commit for Atheme::ChanServ.info: founder, successor, and registered.

Each parse the raw data from Atheme (can be called on from the returned objects #to_raw) and find
the appropriate data.

We've also removed all services files, because right now they seem to be a bunch of files that
serve no purpose other than to have the specific namespaces. Now they're just all created via
module_eval.

Comming up next is refactoring the specs so we do not have to rely on a live IRC/Services setup
to test.